requirements.txt
List of requirements:
pillow
gtts
moviepy